print_proctor_backend/requirements.txt
fastapi == 0.70.0
uvicorn == 0.15.0
motor == 2.4.0
pymongo[srv] == 3.12.0
passlib[bcrypt]
python-multipart
python-jose[cryptography]